requirements.txt

streamlit>=1.32
PyYAML>=6.0
python-docx>=1.0
docxtpl>=0.16
docx2pdf>=0.1.8
pydantic>=2.7
